+++ Neuro/ActivationFunctions/Tanh.cs
@@ -4,21 +4,21 @@ namespace Neuro.ActivationFunctions
 {
     public class Tanh : IActivationFunction
     {
-        public double Alpha { get; set; }
-        public double MinRange { get; set; } = -1;
-        public double MaxRange { get; set; } = 1;
+        public float Alpha { get; set; }
+        public float MinRange { get; set; } = -1;
+        public float MaxRange { get; set; } = 1;
         
-        public double Activation(double x)
+        public float Activation(float x)
         {
-            double e2x = (double)Math.Exp(2 * x);
+            float e2x = (float)Math.Exp(2 * x);
             return (e2x - 1f) / (e2x + 1f);
         }
 
-        public double Derivative(double x)
+        public float Derivative(float x)
         {
-            double
-                eminus2x = (double)Math.Exp(-x),
-                e2x = (double)Math.Exp(x),
+            float
+                eminus2x = (float)Math.Exp(-x),
+                e2x = (float)Math.Exp(x),
                 sum = eminus2x + e2x,
                 square = sum * sum,
                 div = 4 / square;
